+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './extractRawPrivateKeyFromPkcs8';
 export * from './getErrorMessage';
+export * from './resolveCurveName';